improved testing, understand the code properly now
